# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SmartRanks Technical Test API",
        "description": "A comprehensive RESTful API for SmartRanks jobsearch platform. This API provides endpoints for managing products, categories, and user authentication. Built as part of a technical assessment to demonstrate RESTful API development best practices.",
        "contact": {
            "name": "Felipe Murillo",
            "url": "https://github.com/afmurillo97",
            "email": "afmurillo97@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://www.technical-test.site",
            "description": "Production Server"
        },
        {
            "url": "http://api.smartranks.test",
            "description": "Local Development Server"
        },
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/v1/categories": {
            "get": {
                "tags": [
                    "Categories V1"
                ],
                "summary": "List all categories",
                "operationId": "e225c2b7eb5daf7fb16e00f4f07ff030",
                "parameters": [
                    {
                        "name": "include",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "products"
                            ]
                        }
                    },
                    {
                        "name": "category_name[like]",
                        "in": "query",
                        "description": "Filter by partial category name (contains)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "created_at[eq]",
                        "in": "query",
                        "description": "Filter by exact creation date",
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    {
                        "name": "created_at[gt]",
                        "in": "query",
                        "description": "Filter for records created after specified date",
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Categories list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Categories retrieved successfully"
                                        },
                                        "response": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 7
                                                            },
                                                            "category_name": {
                                                                "type": "string",
                                                                "example": "Electronics"
                                                            },
                                                            "description_excerpt": {
                                                                "type": "string",
                                                                "example": "Latest electronic devices"
                                                            },
                                                            "created_at": {
                                                                "type": "string",
                                                                "format": "date-time"
                                                            },
                                                            "url_image": {
                                                                "type": "string",
                                                                "nullable": true
                                                            },
                                                            "products": {
                                                                "type": "array",
                                                                "items": {
                                                                    "$ref": "#/components/schemas/ProductResource"
                                                                },
                                                                "nullable": true
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "meta": {
                                                    "properties": {
                                                        "organization": {
                                                            "type": "string"
                                                        },
                                                        "final_tester": {
                                                            "type": "string"
                                                        },
                                                        "authors": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "string"
                                                            }
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Categories V1"
                ],
                "summary": "Create new category (only Admin)",
                "operationId": "dd76b8d73b7ea8b4951f03d7c0904c92",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreCategoryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Category created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Category created successfully!!"
                                        },
                                        "response": {
                                            "properties": {
                                                "category_id": {
                                                    "type": "integer",
                                                    "example": 5
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden (admin role required)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Forbidden"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid Data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The name field is required."
                                        },
                                        "errors": {
                                            "properties": {
                                                "name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The name field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ProductResource": {
                "title": "Product Resource",
                "description": "Product resource with transformed data",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 7
                    },
                    "name": {
                        "type": "string",
                        "example": "Product Name"
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "category": {
                        "type": "string",
                        "example": "Electronics"
                    },
                    "sku": {
                        "type": "string",
                        "example": "PRD-001"
                    },
                    "description": {
                        "type": "string",
                        "example": "Product description"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 19.99
                    },
                    "stock": {
                        "type": "integer",
                        "example": 100
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00Z"
                    },
                    "product_images": {
                        "properties": {
                            "featured_image": {
                                "type": "string",
                                "example": "http://example.com/image.jpg"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "StoreCategoryRequest": {
                "title": "Request Structure to Store Category",
                "required": [
                    "name",
                    "status"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Electronics"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "featured_image": {
                        "type": "string",
                        "nullable": true
                    },
                    "status": {
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            },
...
            "ErrorResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT token authentication. Include the token in the Authorization header as 'Bearer {token}'",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}